+++ src/Apps/Browser/Messages.elm
@@ -6,8 +6,9 @@ import Apps.Browser.Menu.Messages as Menu
 type Msg
     = MenuMsg Menu.Msg
     | UpdateAddress String
-    | AddressEnter
     | GoPrevious
     | GoNext
     | PageMsg
     | TabGo Int
+    | GoAddress String
+    | NewTabInAddress String
