+++ src/Apps/Browser/Pages/Default/View.elm
@@ -1,8 +1,8 @@
 module Apps.Browser.Pages.Default.View exposing (view)
 
-import Html exposing (Html, div, text)
+import Html exposing (Html, div)
 
 
 view : Html Never
 view =
-    div [] [ text "Soon..." ]
+    div [] []
