+++ src/Apps/Browser/Pages/FBI/View.elm
@@ -1,0 +1,8 @@
+module Apps.Browser.Pages.FBI.View exposing (view)
+
+import Html exposing (Html, div, text)
+
+
+view : Html Never
+view =
+    div [] [ text "TODO" ]
