+++ src/Apps/Browser/Pages/Home/Messages.elm
@@ -1,0 +1,6 @@
+module Apps.Browser.Pages.Home.Messages exposing (Msg(..))
+
+
+type Msg
+    = BrowserGoAddress String
+    | BrowserTabAddress String
