+++ src/Apps/Browser/Pages/Models.elm
@@ -3,13 +3,19 @@ module Apps.Browser.Pages.Models
         ( Model(..)
         , initialModel
         , getSite
-        , getUrl
         , getTitle
         )
 
 import Game.Web.Types as Web
 import Apps.Browser.Pages.NotFound.Models as PageNotFound
 import Apps.Browser.Pages.Default.Models as PageDefault
+import Apps.Browser.Pages.Profile.Models as PageProfile
+import Apps.Browser.Pages.Directory.Models as PageDirectory
+import Apps.Browser.Pages.MissionCenter.Models as PageMissionCenter
+import Apps.Browser.Pages.DownloadCenter.Models as PageDownloadCenter
+import Apps.Browser.Pages.ISP.Models as PageISP
+import Apps.Browser.Pages.FBI.Models as PageFBI
+import Apps.Browser.Pages.News.Models as PageNews
 
 
 type Model
@@ -29,6 +35,7 @@ type Model
     | FBIModel
     | NewsModel
     | BithubModel
+    | MissionCenterModel
 
 
 type alias Meta =
@@ -40,6 +47,9 @@ initialModel ({ type_, meta } as site) =
     case ( type_, meta ) of
         ( Web.Blank, _ ) ->
             BlankModel
+
+        ( Web.Home, _ ) ->
+            HomeModel
 
         ( Web.Unknown, _ ) ->
             UnknownModel
@@ -54,6 +64,27 @@ initialModel ({ type_, meta } as site) =
                 |> PageDefault.initialModel
                 |> DefaultModel
 
+        ( Web.Profile, _ ) ->
+            ProfileModel
+
+        ( Web.Directory, _ ) ->
+            DirectoryModel
+
+        ( Web.MissionCenter, _ ) ->
+            MissionCenterModel
+
+        ( Web.DownloadCenter, _ ) ->
+            DownloadCenterModel
+
+        ( Web.ISP, _ ) ->
+            ISPModel
+
+        ( Web.FBI, _ ) ->
+            FBIModel
+
+        ( Web.News, _ ) ->
+            NewsModel
+
         _ ->
             UnknownModel
 
@@ -67,33 +98,66 @@ getTitle model =
         UnknownModel ->
             "Loading..."
 
+        HomeModel ->
+            "Home"
+
+        ProfileModel ->
+            PageProfile.getTitle
+
+        DirectoryModel ->
+            PageDirectory.getTitle
+
+        MissionCenterModel ->
+            PageMissionCenter.getTitle
+
+        DownloadCenterModel ->
+            PageDownloadCenter.getTitle
+
+        ISPModel ->
+            PageISP.getTitle
+
+        FBIModel ->
+            PageFBI.getTitle
+
+        NewsModel ->
+            PageNews.getTitle
+
         _ ->
             "New Tab"
 
 
-getUrl : Model -> String
-getUrl model =
-    let
-        site =
-            getSite model
-    in
-        site.url
-
-
-getSite : Model -> Web.Site
+getSite : Model -> ( Web.Type, Maybe Web.Meta )
 getSite model =
     case model of
         NotFoundModel model ->
             PageNotFound.getSite model
 
         BlankModel ->
-            { type_ = Web.Blank
-            , url = "about:blank"
-            , meta = Nothing
-            }
+            ( Web.Blank, Nothing )
+
+        HomeModel ->
+            ( Web.Home, Nothing )
+
+        ProfileModel ->
+            PageProfile.getSite
+
+        DirectoryModel ->
+            PageDirectory.getSite
+
+        MissionCenterModel ->
+            PageMissionCenter.getSite
+
+        DownloadCenterModel ->
+            PageDownloadCenter.getSite
+
+        ISPModel ->
+            PageISP.getSite
+
+        FBIModel ->
+            PageFBI.getSite
+
+        NewsModel ->
+            PageNews.getSite
 
         _ ->
-            { type_ = Web.Unknown
-            , url = ""
-            , meta = Nothing
-            }
+            ( Web.Unknown, Nothing )
