+++ src/Apps/Browser/Pages/News/View.elm
@@ -1,0 +1,8 @@
+module Apps.Browser.Pages.News.View exposing (view)
+
+import Html exposing (Html, div, text)
+
+
+view : Html Never
+view =
+    div [] [ text "TODO" ]
