+++ src/Apps/Browser/Pages/NotFound/Models.elm
@@ -27,6 +27,6 @@ getTitle { title } =
     title
 
 
-getSite : Model -> Web.Site
+getSite : Model -> ( Web.Type, Maybe Web.Meta )
 getSite { site } =
-    site
+    ( Web.NotFound, Nothing )
