+++ src/Apps/Hebamp/Models.elm
@@ -59,7 +59,13 @@ initialModel =
 initialHebamp : Hebamp
 initialHebamp =
     { playerId = "player-id"
-    , now = Nothing
+    , now =
+        Just
+            { mediaUrl = "https://upload.wikimedia.org/wikipedia/en/2/2a/Nyan_cat.ogg"
+            , mediaType = "audio/ogg"
+            , label = "Nyan Cat"
+            , duration = 7.4
+            }
     , prev = []
     , next = []
     , currentTime = 0
