+++ src/Game/Web/Models.elm
@@ -11,6 +11,7 @@ module Game.Web.Models
 
 import Dict exposing (Dict)
 import Game.Web.Types exposing (..)
+import Game.Web.Dummy exposing (dummyTunnel)
 
 
 type alias Model =
@@ -40,25 +41,61 @@ isHackable t =
             False
 
 
+blankSite : String -> Site
+blankSite url =
+    { url = url
+    , type_ = Blank
+    , meta = Nothing
+    }
+
+
+unknownSite : String -> Site
+unknownSite url =
+    { url = url
+    , type_ = Unknown
+    , meta = Nothing
+    }
+
+
+homepageSite : String -> Site
+homepageSite url =
+    { url = url
+    , type_ = Default
+    , meta = Nothing
+    }
+
+
 get : String -> Model -> Site
 get str model =
-    case str of
-        "about:blank" ->
-            { url = str
-            , type_ = Blank
-            , meta = Nothing
-            }
+    case String.split "/" str of
+        [ "about:blank" ] ->
+            blankSite str
 
-        str ->
-            case Dict.get str model of
-                Just site ->
-                    site
+        [ "about:home" ] ->
+            homepageSite str
 
-                Nothing ->
+        [ "" ] ->
+            blankSite str
+
+        host :: req ->
+            case List.reverse (String.split "." host) of
+                -- Top Level Domains go here
+                "dmy" :: domain ->
                     { url = str
-                    , type_ = Unknown
+                    , type_ = dummyTunnel <| domain ++ req
                     , meta = Nothing
                     }
+
+                _ ->
+                    case Dict.get str model of
+                        Just site ->
+                            site
+
+                        Nothing ->
+                            unknownSite str
+
+        _ ->
+            unknownSite str
 
 
 refresh : String -> Site -> Model -> Model
