+++ src/Utils/Ports/Map.elm
@@ -1,0 +1,7 @@
+port module Utils.Ports.Map exposing (..)
+
+
+port mapInit : ( String, Float, Float, Int ) -> Cmd msg
+
+
+port mapCenter : ( String, Float, Float, Int ) -> Cmd msg
